+++ nodes/FilterPwd/models/FilterPwdResult.ts
@@ -1,0 +1,5 @@
+
+export class FilterPwdResult {
+	file: string | null = null;
+	count: number | null = null;
+}