+++ CHANGELOG.md
@@ -1,6 +1,18 @@
+## 1.0.2
+
+### Features
+
+* Filter Password Dictionary action
+
+## 1.0.1
+
+### Features
+
+* Hydra One Dictionary Bruteforce (user:password file format)
+
 ## 1.0.0
 
 ### Features
 
 * Hydra Random Bruteforce
-* Hydra Dictionary Bruteforce
+* Hydra Dictionary Bruteforce (users & passwords files)